Subindo refatoração do código da página de signup
